Enable the copying of local files in the container's rootfs.
Users can now define the local files that they want to pass in the
container rootfs in the files field of args.nix. Each file will be
copied inside the container rrotfs in the same path as the value of the
respective entry.

All the files will be entered on the same layer as links and later
another layer will add the actual files. This is typical for Nix.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>