Use filesets for better control over the copying files
Nix allows to define the exact files that we are interested in to copy
in the Nix store. We can do that with filesets. The advantage is that a
change in the local directory will not affect the derivation and
therefore, we will not need to rebuild it. Only changes to the specified
files will result to rebuilding the derivation.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>