Change the creation time of the container and the compressor
By default Nix uses "1970-01-01T00:00:01Z" as the creation time. We
chnage tat to use the local time, however, as described in Nix
documentation, this affects the reproducability of the image. On the
other hand, we do not really care about it for the time being.

In addition we change the compressor to none, so we can have a tar
output that can be imported in ctr directly.

Signed-off-by: Charalampos Mainas <cmainas@nubificus.co.uk>